refactor: restructure application by separating app, router, and server logic; update Vercel configuration (#30)
